@@ -14,50 +14,49 @@
     "start:emulators-import": "firebase emulators:start --only auth,functions,firestore,pubsub --import ./firestoreData",
     "test:e2e": "cross-env FIRESTORE_EMULATOR_HOST=\"localhost:8080\" FIREBASE_AUTH_EMULATOR_HOST=\"localhost:9099\" FIREBASE_DATABASE_EMULATOR_HOST=\"localhost:9000\" E2E_TEST=true start-test \"quasar dev\" http-get://localhost:5000 \"cypress open\"",
     "test:e2e:ci": "cross-env E2E_TEST=true start-test \"quasar dev\" http-get://localhost:5000 \"cypress run\"",
-    "test:e2e:emulated": "firebase emulators:exec --only auth,functions,firestore,database \"npm run test:e2e\"",
-    "test:rules:fire": "firebase emulators:exec --only firestore \"jest firestore.rules._jest.spec.js --detectOpenHandles\"",
-    "test:rules:rtdb": "firebase emulators:exec --only database \"jest database.rules._jest.spec.js --detectOpenHandles\"",
-    "test:rules:fire:watch": "firebase emulators:exec --only firestore \"jest firestore.rules._jest.spec.js --detectOpenHandles --watchAll\"",
-    "test:rules:rtdb:watch": "firebase emulators:exec --only database \"jest database.rules._jest.spec.js --detectOpenHandles --watchAll\"",
+    "test:e2e:emulated": "firebase emulators:exec --only auth,functions,firestore,database --ui \"npm run test:e2e\"",
+    "test:e2e:ci:emulated": "firebase emulators:exec --only auth,functions,firestore,database --ui \"npm run test:e2e:ci\"",
     "test:jest": "jest --watchAll --verbose=false",
     "test:functions:emulated": "firebase emulators:exec --only auth,functions,firestore,database \"jest --watchAll\""
   },
   "dependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.16.8",
-    "@babel/preset-env": "^7.16.8",
-    "@firebase/rules-unit-testing": "^2.0.1",
-    "@quasar/extras": "^1.12.3",
-    "babel-jest": "^27.2.4",
+    "@babel/preset-env": "^7.16.11",
+    "@firebase/rules-unit-testing": "^2.0.2",
+    "@quasar/extras": "^1.12.5",
+    "babel-jest": "^27.5.0",
     "chart.js": "^3.6.0",
-    "eslint-plugin-vue": "^8.3.0",
+    "core-js": "^3.21.0",
     "esm": "^3.2.25",
     "file-saver": "^2.0.5",
-    "firebase": "^9.6.3",
-    "quasar": "^2.4.10",
+    "firebase": "^9.6.6",
+    "firebase-functions": "^3.17.1",
+    "quasar": "^2.5.3",
     "ts-jest": "^27.1.3",
+    "vue": "^3.2.29",
+    "vue-router": "^4.0.12",
     "vuedraggable": "^4.1.0",
-    "core-js": "^3.6.5",
     "vuex": "^4.0.1"
   },
   "devDependencies": {
-    "@quasar/app": "^3.1.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@quasar/app": "^3.3.2",
     "@quasar/quasar-app-extension-qcalendar": "^4.0.0-beta.3",
-    "@quasar/quasar-app-extension-testing-e2e-cypress": "^3.1.0",
+    "@quasar/quasar-app-extension-testing-e2e-cypress": "^4.0.1",
+    "@types/node": "^17.0.15",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
     "cross-env": "^7.0.3",
-    "cypress": "^9.2.1",
+    "cypress": "^9.4.1",
     "cypress-firebase": "^2.2.1",
+    "eslint": "^8.8.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "firebase-admin": "^10.0.0",
+    "eslint-plugin-vue": "^8.4.1",
+    "firebase-admin": "^10.0.2",
     "stylus": "^0.56.0",
     "stylus-loader": "^6.1.0",
-    "workbox-webpack-plugin": "^6.2.2",
-    "@babel/eslint-parser": "^7.13.14",
-    "@types/node": "^12.20.21",
-    "@typescript-eslint/eslint-plugin": "^4.16.1",
-    "@typescript-eslint/parser": "^4.16.1",
-    "eslint": "^7.14.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-vue": "^7.0.0"
+    "workbox-webpack-plugin": "^6.2.2"
   },
   "browserslist": [
     "last 10 Chrome versions",
