@@ -27,17 +27,18 @@
 </template>
 
 <script>
-import { useStore } from "vuex";
-import { useQuasar } from "quasar";
-import { ref, computed, defineAsyncComponent } from "vue";
-import { deleteDoc, doc, getFirestore } from "@firebase/firestore";
+import { useStore } from 'vuex';
+import { useQuasar } from 'quasar';
+import { ref, computed, defineAsyncComponent, watch } from 'vue';
+import { deleteDoc, doc, getFirestore } from '@firebase/firestore';
 export default {
-  name: "Delete-Tree",
+  name: 'Delete-Tree',
   props: {
     menu: Boolean,
     treeOpt: { label: String, id: String, tree: [] },
   },
-  setup(props) {
+  emits: ['open'],
+  setup(props, { emit }) {
     const q = useQuasar();
     const store = useStore();
     const confirm = ref(false);
@@ -49,26 +50,29 @@ export default {
         doc(getFirestore(), `/movements/${movement.value.id}/trees/${id}`)
       )
         .then((res) => {
-          store.commit("movement/removeTree", id);
+          store.commit('movement/removeTree', id);
           return q.notify({
-            color: "positive",
-            textColor: "white",
-            icon: "cloud_download",
-            message: "Tree Deleted",
+            color: 'positive',
+            textColor: 'white',
+            icon: 'cloud_download',
+            message: 'Tree Deleted',
           });
         })
         .catch((err) => {
           console.error(err);
           q.notify({
-            color: "negative",
-            textColor: "white",
-            icon: "error",
-            message: "Oops, Something went wrong!",
+            color: 'negative',
+            textColor: 'white',
+            icon: 'error',
+            message: 'Oops, Something went wrong!',
           });
         });
     }
     const movement = computed(() => store.state.movement.movement);
     const user = computed(() => store.state.auth.user);
+    watch(confirm, () => {
+      emit('open', confirm.value);
+    });
     return {
       q,
       confirm,
@@ -77,12 +81,12 @@ export default {
       user,
       removeLoading,
       deleteDialog,
-      treeOpt: props.treeOpt ? props.treeOpt : { label: "" },
+      treeOpt: props.treeOpt ? props.treeOpt : { label: '' },
     };
   },
   components: {
-    "mt-remove-dialog": defineAsyncComponent(() =>
-      import("./mt-remove-dialog.vue")
+    'mt-remove-dialog': defineAsyncComponent(() =>
+      import('./mt-remove-dialog.vue')
     ),
   },
 };
