@@ -4,38 +4,36 @@
       <q-item>
         <q-item-section avatar>Name</q-item-section>
         <q-item-section>
-          <q-item-label>{{ localMember.name }}</q-item-label>
+          <q-item-label>{{ l_member.name }}</q-item-label>
         </q-item-section>
       </q-item>
       <q-item>
         <q-item-section avatar>Role</q-item-section>
         <q-item-section>
-          <q-item-label v-if="localMember.role">
+          <q-item-label v-if="l_member.role">
             {{
-              localMember.role
-                ? roles[localMember.role]
-                  ? roles[localMember.role].label
-                  : localMember.role.label
-                  ? localMember.role.label
-                  : "Missing Role"
-                : "Missing Role"
+              l_member.role &&
+              roles[l_member.role] &&
+              roles[l_member.role].label > ''
+                ? roles[l_member.role].label
+                : 'Missing Role'
             }}
             <q-tooltip
               class="bg-accent text-grey-10"
               anchor="center right"
               self="center left"
               v-if="
-                localMember.role &&
-                (roles[localMember.role] || localMember.role.desc > '')
+                l_member.role &&
+                (roles[l_member.role] || l_member.role.desc > '')
               "
               >{{
-                localMember.role
-                  ? roles[localMember.role]
-                    ? roles[localMember.role].desc
-                    : localMember.role.desc
-                    ? localMember.role.desc
-                    : ""
-                  : ""
+                l_member.role
+                  ? roles[l_member.role]
+                    ? roles[l_member.role].desc
+                    : l_member.role.desc
+                    ? l_member.role.desc
+                    : ''
+                  : ''
               }}</q-tooltip
             >
           </q-item-label>
@@ -44,11 +42,11 @@
       <q-item>
         <q-item-section avatar>Modifiers</q-item-section>
         <q-item-section>
-          <q-item-label v-if="localMember.mods">
+          <q-item-label v-if="l_member.mods">
             <div class="q-gutter-xs">
               <q-badge
                 color="blue"
-                v-for="overrideStyle in localMember.mods"
+                v-for="overrideStyle in l_member.mods"
                 :key="overrideStyle.label"
                 multi-line
               >
@@ -58,8 +56,8 @@
                       ? mods[overrideStyle].label
                       : overrideStyle.label
                       ? overrideStyle.label
-                      : "Missing Modifier"
-                    : "Missing Modifier"
+                      : 'Missing Modifier'
+                    : 'Missing Modifier'
                 }}
                 <q-tooltip
                   class="bg-accent text-grey-10"
@@ -75,8 +73,8 @@
                         ? mods[overrideStyle].desc
                         : overrideStyle.desc
                         ? overrideStyle.desc
-                        : ""
-                      : ""
+                        : ''
+                      : ''
                   }}</q-tooltip
                 >
               </q-badge>
@@ -87,16 +85,16 @@
       <!-- <q-item>
         <q-item-section avatar>Parents</q-item-section>
         <q-item-section>
-          <q-item-label>{{ parents[localMember.parent].name }}</q-item-label>
-          <q-item-label v-if="localMember.alt > ''">{{
-            parents[localMember.alt].name
+          <q-item-label>{{ parents[l_member.parent].name }}</q-item-label>
+          <q-item-label v-if="l_member.alt > ''">{{
+            parents[l_member.alt].name
           }}</q-item-label>
         </q-item-section>
       </q-item> -->
       <q-item>
         <q-item-section avatar>Notes</q-item-section>
         <q-item-section>
-          <q-item-label>{{ localMember.notes }}</q-item-label>
+          <q-item-label>{{ l_member.notes }}</q-item-label>
         </q-item-section>
       </q-item>
     </q-list>
@@ -104,18 +102,18 @@
 </template>
 
 <script>
-import { mapState } from "vuex";
+import { mapState } from 'vuex';
 
 export default {
   // name: 'ComponentName',
   props: {
-    localMember: {},
+    l_member: {},
   },
   data() {
     return {};
   },
   computed: {
-    ...mapState("movement", ["roles", "mods", "parents"]),
+    ...mapState('movement', ['roles', 'mods', 'parents']),
   },
 };
 </script>
