@@ -114,6 +114,7 @@
             data-cy="label"
           >
             <q-btn
+              :disable="!permissions.settings_calc_edit"
               :key="props.row.value"
               style="background-color: white; color: black; width: 100%"
               dense
@@ -149,9 +150,23 @@
                   : JSON.parse(JSON.stringify(defaultCond))
               "
               :styles="styleOptions"
-              @add="onAdd($event, props.row)"
-              @change="onChange($event, props.row)"
-              @remove="onRemove($event, props.row)"
+              @add="
+                ($event) => {
+                  if (permissions.settings_calc_edit) onAdd($event, props.row);
+                }
+              "
+              @change="
+                ($event) => {
+                  if (permissions.settings_calc_edit)
+                    onChange($event, props.row);
+                }
+              "
+              @remove="
+                ($event) => {
+                  if (permissions.settings_calc_edit)
+                    onRemove($event, props.row);
+                }
+              "
             />
           </q-td>
           <q-td
@@ -178,34 +193,34 @@
 </template>
 
 <script>
-import { mapGetters, mapMutations, mapState } from "vuex";
-import { debounce, Notify, LocalStorage, Dark } from "quasar";
-import { updateStyleByKey } from "./../scripts/styles.js";
-import { defineAsyncComponent } from "vue";
+import { mapGetters, mapMutations, mapState } from 'vuex';
+import { debounce, Notify, LocalStorage, Dark } from 'quasar';
+import { updateStyleByKey } from './../scripts/styles.js';
+import { defineAsyncComponent } from 'vue';
 
 const columns = [
   {
-    name: "label",
-    align: "center",
-    label: "Label",
-    field: "label",
+    name: 'label',
+    align: 'center',
+    label: 'Label',
+    field: 'label',
     sortable: true,
-    help: "The name of this statistic",
+    help: 'The name of this statistic',
   },
   {
-    name: "desc",
-    align: "left",
-    label: "Description",
-    field: "desc",
+    name: 'desc',
+    align: 'left',
+    label: 'Description',
+    field: 'desc',
     sortable: true,
-    help: "A short description of what this statistic is",
+    help: 'A short description of what this statistic is',
   },
   {
-    name: "target",
-    align: "left",
-    label: "Target",
-    field: "target",
-    help: "A goal for the value of this statistic",
+    name: 'target',
+    align: 'left',
+    label: 'Target',
+    field: 'target',
+    help: 'A goal for the value of this statistic',
   },
   // {
   //   name: "current",
@@ -215,18 +230,18 @@ const columns = [
   //   help: "The current value of this statistic (Tree Total | No Parent Total)",
   // },
   {
-    name: "condition",
-    label: "Conditions",
-    field: "condition",
-    align: "left",
-    help: "Define a theoretical member of the Movement to search for",
+    name: 'condition',
+    label: 'Conditions',
+    field: 'condition',
+    align: 'left',
+    help: 'Define a theoretical member of the Movement to search for',
   },
   {
-    name: "unit",
-    label: "Unit",
-    field: "unit",
-    align: "center",
-    help: "The unit for this statistic",
+    name: 'unit',
+    label: 'Unit',
+    field: 'unit',
+    align: 'center',
+    help: 'The unit for this statistic',
   },
 ];
 export default {
@@ -237,24 +252,24 @@ export default {
       isFullscreen: false,
       hidebaseStyles: true,
       pagination: {
-        sortBy: "label",
+        sortBy: 'label',
         descending: false,
         page: 1,
         rowsPerPage: 10,
       },
       defaultCond: {
-        class: "expression",
+        class: 'expression',
         elements: [
           {
-            class: "condition",
-            id: "",
+            class: 'condition',
+            id: '',
           },
           {
-            class: "condition",
-            id: "",
+            class: 'condition',
+            id: '',
           },
         ],
-        operator: "plus",
+        operator: 'plus',
       },
     };
   },
@@ -262,66 +277,66 @@ export default {
     this.Dark = Dark;
     this.updateStyle = debounce(this.updateStyle, 2000);
     // this.updateMovement = debounce(this.updateMovement, 2000)
-    this.pagination = LocalStorage.has("movementEditPagination")
-      ? LocalStorage.getItem("movementEditPagination")
+    this.pagination = LocalStorage.has('movementEditPagination')
+      ? LocalStorage.getItem('movementEditPagination')
       : {
-          sortBy: "label",
+          sortBy: 'label',
           descending: false,
           page: 1,
           rowsPerPage: 10,
         };
   },
   methods: {
-    ...mapMutations("movement", ["setStyleLoading"]),
+    ...mapMutations('movement', ['setStyleLoading']),
     updateStyle(id, key, val) {
       this.setStyleLoading({ id, val: true });
       return updateStyleByKey(this.$route.params.movId, id, key, val)
         .then(() => {
           this.setStyleLoading({ id, val: true });
           return Notify.create({
-            color: "positive",
-            textColor: "white",
-            icon: "cloud_download",
-            message: "Style Updated",
+            color: 'positive',
+            textColor: 'white',
+            icon: 'cloud_download',
+            message: 'Style Updated',
           });
         })
         .catch((err) => {
           this.setStyleLoading({ id, val: false });
           Notify.create({
-            color: "negative",
-            textColor: "white",
-            icon: "error",
-            message: "Oops, Something went wrong!",
+            color: 'negative',
+            textColor: 'white',
+            icon: 'error',
+            message: 'Oops, Something went wrong!',
           });
         });
     },
     onAdd($event, prop) {
       // console.log($event, prop);
-      if ($event.el.class === "expression") {
-        if ($event.type === "condition") {
+      if ($event.el.class === 'expression') {
+        if ($event.type === 'condition') {
           $event.el.elements.push({
-            class: "condition",
-            id: "",
+            class: 'condition',
+            id: '',
           });
-        } else if ($event.type === "number") {
+        } else if ($event.type === 'number') {
           $event.el.elements.push({
-            class: "number",
+            class: 'number',
             value: 0,
           });
-        } else if ($event.type === "expression") {
+        } else if ($event.type === 'expression') {
           $event.el.elements.push({
-            class: "expression",
+            class: 'expression',
             elements: [
               {
-                class: "condition",
-                id: "",
+                class: 'condition',
+                id: '',
               },
               {
-                class: "condition",
-                id: "",
+                class: 'condition',
+                id: '',
               },
             ],
-            operator: "plus",
+            operator: 'plus',
           });
         }
       }
@@ -329,7 +344,7 @@ export default {
       return updateStyleByKey(
         this.$route.params.movId,
         prop.id,
-        "condition",
+        'condition',
         prop.condition
       );
     },
@@ -343,12 +358,12 @@ export default {
       return updateStyleByKey(
         this.$route.params.movId,
         prop.id,
-        "condition",
+        'condition',
         prop.condition
       );
     },
     onRemove(path, prop) {
-      function resolve(path, obj = self, separator = ".") {
+      function resolve(path, obj = self, separator = '.') {
         var properties = Array.isArray(path) ? path : path.split(separator);
         return properties.reduce((prev, curr) => prev && prev[curr], obj);
       }
@@ -362,13 +377,13 @@ export default {
       el.splice(index, 1);
       if (el.length === 1) {
         const temp = { ...el[0] };
-        let shortPath = arrayPath.substring(0, arrayPath.lastIndexOf("."));
+        let shortPath = arrayPath.substring(0, arrayPath.lastIndexOf('.'));
         // console.log(shortPath);
         const shortIndex = shortPath.substring(
-          shortPath.lastIndexOf(".") + 1,
-          arrayPath.lastIndexOf(".")
+          shortPath.lastIndexOf('.') + 1,
+          arrayPath.lastIndexOf('.')
         );
-        shortPath = shortPath.substring(0, shortPath.lastIndexOf("."));
+        shortPath = shortPath.substring(0, shortPath.lastIndexOf('.'));
         // console.log(shortIndex);
         let parentEl = resolve(shortPath, prop.condition);
         if (Array.isArray(parentEl)) {
@@ -378,30 +393,30 @@ export default {
       return updateStyleByKey(
         this.$route.params.movId,
         prop.id,
-        "condition",
+        'condition',
         prop.condition
       );
     },
     evalCalculated(cond, stats, totalId) {
       if (!(cond && stats)) return 0;
       const values = [];
-      let total = totalId ? totalId : "total";
-      if (cond.class === "expression") {
+      let total = totalId ? totalId : 'total';
+      if (cond.class === 'expression') {
         for (var el in cond.elements) {
           values.push(this.evalCalculated(cond.elements[el], stats, total));
         }
         if (values <= []) {
           return 0;
-        } else if (cond.operator === "plus") {
+        } else if (cond.operator === 'plus') {
           return values.reduce((a, b) => a + b, 0);
-        } else if (cond.operator === "minus") {
+        } else if (cond.operator === 'minus') {
           return values.reduce((a, b) => a - b, 0);
-        } else if (cond.operator === "multiply") {
+        } else if (cond.operator === 'multiply') {
           return values.reduce((a, b) => a * b);
-        } else if (cond.operator === "divide") {
+        } else if (cond.operator === 'divide') {
           return values.reduce((a, b) => a / b);
         }
-      } else if (cond.class === "condition") {
+      } else if (cond.class === 'condition') {
         return parseInt(
           stats[cond.id]
             ? stats[cond.id][total]
@@ -411,7 +426,7 @@ export default {
               : 0
             : 0
         );
-      } else if (cond.class === "number") {
+      } else if (cond.class === 'number') {
         return parseFloat(cond.value);
       }
 
@@ -419,21 +434,21 @@ export default {
     },
   },
   computed: {
-    ...mapGetters("movement", [
-      "roleOpts",
-      "modOpts",
-      "complexOpts",
-      "calcOpts",
-      "backgroundColor",
-      "color",
+    ...mapGetters('movement', [
+      'roleOpts',
+      'modOpts',
+      'complexOpts',
+      'calcOpts',
+      'backgroundColor',
+      'color',
     ]),
-    ...mapState("movement", [
-      "movement",
-      "roles",
-      "mods",
-      "totalStats",
-      "permissions",
-      "stats",
+    ...mapState('movement', [
+      'movement',
+      'roles',
+      'mods',
+      'totalStats',
+      'permissions',
+      'stats',
     ]),
     styleOptions() {
       // console.log(
@@ -445,7 +460,11 @@ export default {
       return this.roleOpts
         .concat(this.modOpts)
         .concat(this.complexOpts)
-        .concat(Object.values(this.totalStats ? this.totalStats : {}))
+        .concat({
+          label: 'Tree Total',
+          id: 'treeTotal',
+          desc: 'The number of Members on the tree',
+        })
         .sort((a, b) => {
           return a.label > b.label ? 1 : a.label < b.label ? -1 : 0;
         });
@@ -471,15 +490,15 @@ export default {
   },
   watch: {
     pagination() {
-      LocalStorage.set("movementEditPagination", this.pagination);
+      LocalStorage.set('movementEditPagination', this.pagination);
     },
   },
   components: {
-    "mt-style-menu": defineAsyncComponent(() =>
-      import("./../components/actions/mt-style-menu.vue")
+    'mt-style-menu': defineAsyncComponent(() =>
+      import('./../components/actions/mt-style-menu.vue')
     ),
-    "mt-calculated-node": defineAsyncComponent(() =>
-      import("./mt-calculated-node.vue")
+    'mt-calculated-node': defineAsyncComponent(() =>
+      import('./mt-calculated-node.vue')
     ),
   },
 };
