@@ -1,10 +1,13 @@
 <template>
   <div>
     <mt-movement-toolbox class="print-hide" :treeOpt="visibleTree" />
-    <mt-filter-sort :tab="tab" :treeOpt="visibleTree" />
+    <mt-filter-sort :tab="tab" />
     <q-scroll-area style="height: calc(100vh - 50px); max-width: 100%">
       <div style="padding-left: 56px; min-height: 90vh" data-cy="background">
         <MovBanner />
+        <div class="print-only text-center">
+          {{ tab === 'trees' ? visibleTree.label : '' }}
+        </div>
         <q-tabs
           v-model="tab"
           dense
@@ -13,6 +16,7 @@
           inline-label
           align="left"
           data-cy="member-tabs"
+          class="print-hide"
         >
           <q-tab
             name="trees"
@@ -35,7 +39,7 @@
               option-value="id"
             >
               <template v-slot:selected>
-                <div class="text-black">
+                <div :style="`color:${color}`">
                   {{ visibleTree ? visibleTree.label : '' }}
                 </div>
               </template>
@@ -126,7 +130,6 @@ export default {
     watch(
       visibleTree,
       (newVal, oldVal) => {
-        console.log(oldVal, newVal, visibleTree.value);
         if (
           (!oldVal && newVal) ||
           (oldVal && newVal && newVal.id !== oldVal.id)
